Set default method when selecting service
This is better UX. Instead of always forcing the user to pick a method
to call, we place the first one in there on the off-chance that it's the
one they wanted.

I'll need to make the reducer handle most of this logic instead of in
the Editor component. Kinda confused because I've got two reducers, the
one that handles the selected service state knows nothing about the
registry, making it hard for me to figure out what method to pick.